backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
requests==2.31.0
djangorestframework-simplejwt>=5.2.2  
django-filter==25.1
python-decouple==3.7
django-cors-headers==3.10.0
shortuuid==1.0.9
webcolors==1.11.1
django-bootstrap5==22.2
pytest-pythonpath==0.7.3
pytest-django==4.4.0
drf-yasg==1.21
reportlab==4.2.5
flake8
drf-extra-fields==3.7.0
dotenv